refactor: Controller, Service 헬퍼 통합
