+++ backend/src/test/java/feedzupzup/backend/config/E2EHelper.java
@@ -2,24 +2,21 @@ package feedzupzup.backend.config;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@TestcontainersTest
-public abstract class E2EHelper {
+/**
+ * E2E (End-to-End) 테스트용 헬퍼 클래스
+ *
+ * RestAssured를 사용한 HTTP 요청 테스트를 위한 설정을 제공합니다.
+ * IntegrationTestSupport를 상속받아 Context를 공유합니다.
+ */
+public abstract class E2EHelper extends IntegrationTestSupport {
 
     @LocalServerPort
     private int port;
 
-    @Autowired
-    private DataInitializer dataInitializer;
-
     @BeforeEach
-    void setUp() {
-        dataInitializer.deleteAll();
+    void setUpRestAssured() {
         RestAssured.port = port;
     }
 }
